create main stump
