+++ matx/src/app/views/dashboard/Analytics.jsx
@@ -30,29 +30,9 @@ const H4 = styled('h4')(({ theme }) => ({
   color: theme.palette.text.secondary,
 }));
 
-const Analytics = () => {
+const Analytics = ({videos}) => {
   const { palette } = useTheme();
-  const [videos, setVideos] = useState([])
-  async function get_data() {
-    let hasMore = true
-    let cursor = "0"
-    let vids = [];
-
-    while  (hasMore) {
-      const result = await fetch("http://127.0.0.1:8000/tonedetection/getvideos?" + new URLSearchParams({cursor}));
-      const json = await result.json()
-      vids = [...vids, ...json.data.videos]
-      const sortedVideos = vids.sort((a, b) => b.create_time - a.create_time)
-      setVideos([...sortedVideos]); 
-      hasMore = Boolean(json.data.hasMore)
-      cursor = json.data.cursor
-    } 
-
-  }
-
-  useEffect(() => {
-    get_data()
-  }, [])
+  
   return (
     <Fragment>
       <ContentBox className="analytics">
